wip, refactor done, todo: fix issue parsing/validating
