+++ .gitignore
@@ -6,3 +6,5 @@ node_modules
 *.iml
 out
 gen
+
+dist
