+++ README.md
@@ -8,8 +8,23 @@ It uses [adapter.js](https://github.com/webrtc/adapter) for WebRTC and [socket.i
 ### Dependencies
 
 * [NodeJS](https://nodejs.org)
+* [Docker](https://www.docker.com)
 
-### Starting and debugging
+### Starting with Docker
+
+Go to the directory that has your Dockerfile and run the following command to build the Docker image. The -t flag lets you tag your image so it's easier to find later using the docker images command:
+
+```
+docker build . -t <your username>/webrtc-app
+```
+
+Run the image you previously built:
+
+```
+docker run -p 8080:80 -d <your username>/webrtc-app
+```
+
+### Starting and debugging (without docker)
 
 For debugging all logs:
