+++ package.json
@@ -2,12 +2,32 @@
   "name": "native-webrtc-app",
   "version": "0.0.1",
   "description": "WebRTC App",
+  "engines": {
+    "node": ">=8 <16"
+  },
   "dependencies": {
-    "async": ">= 1.5.2",
-    "colors": ">= 1.1.2",
-    "debug": ">= 2.2.0",
-    "express": ">= 4.12.3",
-    "websocket": ">= 1.0.18",
-    "socket.io": ">= 1.3.5"
+    "async": ">=3.2.2",
+    "colors": ">=1.1.2",
+    "debug": ">=4.3.0",
+    "express": "^4.17.2",
+    "socket.io": "^4.4.1"
+  },
+  "devDependencies": {
+    "@types/async": "^3.2.12",
+    "@types/debug": "^4.1.7",
+    "@types/express": "^4.17.13",
+    "@types/node": "^16.11.21",
+    "rimraf": "^3.0.2",
+    "ts-node": "^10.4.0",
+    "ts-node-dev": "^1.1.8",
+    "typescript": "^4.5.5"
+  },
+  "scripts": {
+    "prebuild": "rimraf dist",
+    "build": "tsc",
+    "dev": "tsnd --transpile-only --files src/index.ts",
+    "debug": "tsnd --transpile-only --inspect --files src/index.ts",
+    "cli": "tsnd --transpile-only --files src/index.ts --cli",
+    "test": "jest"
   }
 }
